Expose PartialDateString type
