+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alveusgg/data",
-  "version": "0.17.3",
+  "version": "0.17.4",
   "private": true,
   "license": "SEE LICENSE IN LICENSE.md",
   "repository": {
